requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic
google-genai
gspread
oauth2client
requests
python-multipart
